% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab_test-data.R
\docType{data}
\name{ab_test}
\alias{ab_test}
\title{ab_test}
\format{
An object of class \code{data.frame} with 294478 rows and 2 columns.
}
\usage{
data(ab_test)
}
\description{
50 coin tosses
}
\examples{
data(ab_test)
}
\keyword{datasets}
