% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airline_stats-data.R
\docType{data}
\name{airline_stats}
\alias{airline_stats}
\title{airline_stats}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 33468 rows and 4 columns.
}
\usage{
data(airline_stats)
}
\description{
contains the tax-assessed values for residential properties in King County
}
\examples{
data(airline_stats)
}
\keyword{datasets}
