% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_automobile-data.R
\docType{data}
\name{automobile}
\alias{automobile}
\title{automobile}
\format{
An object of class \code{data.frame} with 201 rows and 29 columns.
}
\usage{
data(automobile)
}
\description{
automobile data for eda demo
}
\examples{
data(automobile)
}
\keyword{datasets}
