% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/click_rate-data.R, R/imanishi-data.R
\docType{data}
\name{click_rates}
\alias{click_rates}
\title{click_rates}
\format{
An object of class \code{data.frame} with 6 rows and 3 columns.

An object of class \code{data.frame} with 6 rows and 3 columns.
}
\usage{
data(click_rates)

data(click_rates)
}
\description{
Headline Click Rates

Digit Frequency Counts of laboratory data
}
\examples{
data(click_rates)
data(click_rates)
}
\keyword{datasets}
