% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfw-data.R
\docType{data}
\name{dfw}
\alias{dfw}
\title{DFW}
\format{
An object of class \code{data.frame} with 1 rows and 5 columns.
}
\usage{
data(dfw)
}
\description{
data set containing population and murder
rates (in units of murders per 100,000 people per year) for each state
}
\examples{
data(dfw)
}
\keyword{datasets}
