% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/four_sessions-data.R
\docType{data}
\name{four_sessions}
\alias{four_sessions}
\title{four_sessions}
\format{
An object of class \code{data.frame} with 20 rows and 2 columns.
}
\usage{
data(four_sessions)
}
\description{
Web Page Session Time data
}
\examples{
data(four_sessions)
}
\keyword{datasets}
