% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heads-data.R
\docType{data}
\name{heads}
\alias{heads}
\title{heads}
\format{
An object of class \code{integer} of length 1.
}
\usage{
data(heads)
}
\description{
50 coin tosses
}
\examples{
data(heads)
}
\keyword{datasets}
