% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states-data.R
\docType{data}
\name{house}
\alias{house}
\title{King County Housing Data}
\format{
An object of class \code{data.frame} with 22689 rows and 22 columns.
}
\usage{
data(house)
}
\description{
housing data from King County (Seattle), Washington
}
\examples{
data(house)
}
\keyword{datasets}
