% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kc_tax-data.R
\docType{data}
\name{kc_tax}
\alias{kc_tax}
\title{kc_tax}
\format{
An object of class \code{data.frame} with 498249 rows and 3 columns.
}
\usage{
data(kc_tax)
}
\description{
contains the tax-assessed values for residential properties in King County
}
\examples{
data(kc_tax)
}
\keyword{datasets}
