% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laptops-data.R
\docType{data}
\name{laptops}
\alias{laptops}
\title{laptops}
\format{
An object of class \code{data.frame} with 1303 rows and 13 columns.
}
\usage{
data(laptops)
}
\description{
data of laptop manifacturers
}
\examples{
data(laptops)
}
\keyword{datasets}
