% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laptops_unclean-data.R
\docType{data}
\name{laptops_unclean}
\alias{laptops_unclean}
\title{laptops_unclean}
\format{
An object of class \code{data.frame} with 1303 rows and 13 columns.
}
\usage{
data(laptops_unclean)
}
\description{
data of laptop manifacturers
}
\examples{
data(laptops_unclean)
}
\keyword{datasets}
