% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lc_loans-data.R
\docType{data}
\name{lc_loans}
\alias{lc_loans}
\title{lc_loans}
\format{
An object of class \code{data.frame} with 450961 rows and 2 columns.
}
\usage{
data(lc_loans)
}
\description{
This is taken from data provided by Lending
Club, a leader in the peer-to-peer lending business. The grade goes from A (high) to
G (low). The outcome is either paid off, current, late, or charged off (the balance of
the loan is not expected to be collected).
}
\examples{
data(lc_loans)
}
\keyword{datasets}
