% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loans3000-data.R
\docType{data}
\name{loan3000}
\alias{loan3000}
\title{loan3000}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3000 rows and 5 columns.
}
\usage{
data(loan3000)
}
\description{
3000 samples of loan data
}
\examples{
data(loan3000)
}
\keyword{datasets}
