% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loan_data-data.R
\docType{data}
\name{loan_data}
\alias{loan_data}
\title{loan_data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 45342 rows and 21 columns.
}
\usage{
data(loan_data)
}
\description{
3000 samples of loan data
}
\examples{
data(loan_data)
}
\keyword{datasets}
