% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loans_income-data.R
\docType{data}
\name{loans_income}
\alias{loans_income}
\title{loans_income}
\format{
An object of class \code{data.frame} with 50000 rows and 1 columns.
}
\usage{
data(loans_income)
}
\description{
This is taken from data provided by Lending
Club, a leader in the peer-to-peer lending business. The grade goes from A (high) to
G (low). The outcome is either paid off, current, late, or charged off (the balance of
the loan is not expected to be collected).
}
\examples{
data(loans_income)
}
\keyword{datasets}
