% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_times-data.R
\docType{data}
\name{session_times}
\alias{session_times}
\title{session_times}
\format{
An object of class \code{data.frame} with 36 rows and 2 columns.
}
\usage{
data(session_times)
}
\description{
Web Page Session Time data
}
\examples{
data(session_times)
}
\keyword{datasets}
