% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp500-data.R
\docType{data}
\name{sp500_px}
\alias{sp500_px}
\title{SP500_px}
\format{
An object of class \code{data.frame} with 5647 rows and 518 columns.
}
\usage{
data(sp500_px)
}
\description{
the daily returns for telecommunication stocks
from July 2012 through June 2015.
}
\examples{
data(sp500_px)
}
\keyword{datasets}
