% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp500-data.R
\docType{data}
\name{sp500_sym}
\alias{sp500_sym}
\title{SP500_sym}
\format{
An object of class \code{data.frame} with 517 rows and 4 columns.
}
\usage{
data(sp500_sym)
}
\description{
the daily returns for telecommunication stocks
from July 2012 through June 2015.
}
\examples{
data(sp500_sym)
}
\keyword{datasets}
