% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/house-data.R
\docType{data}
\name{state}
\alias{state}
\title{State}
\format{
An object of class \code{data.frame} with 50 rows and 4 columns.
}
\usage{
data(state)
}
\description{
State data for Murder Rates with Abbrev.
}
\examples{
data(state)
}
\keyword{datasets}
