% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weatherAUS-data.R
\docType{data}
\name{weatherAUS}
\alias{weatherAUS}
\title{Weather Data from Australia}
\format{
An object of class \code{data.frame} with 142193 rows and 24 columns.
}
\usage{
data(weatherAUS)
}
\description{
Weather Data from Australia
}
\examples{
data(weatherAUS)
}
\keyword{datasets}
